src/sorting_functions/quick-sort.ts
function quickSort(arr: number[]): number[] {
  if (arr.length <= 1) {
    return arr;
  }

  const midIndex = Math.floor(arr.length / 2);
  const pivot = arr[midIndex];
  const left = [];
  const right = [];

  for (let i = 0; i < arr.length; i++) {
    if (i === midIndex) continue; // Skip the pivot itself
    if (arr[i] < pivot) {
      left.push(arr[i]);
    } else {
      right.push(arr[i]);
    }
  }

  return quickSort(left).concat(pivot, quickSort(right));
}

console.log(quickSort([5, 3, 7, 6, 2, 9]));

export default quickSort;
